shift API tests into integration
